api/requirements.txt
python-dotenv==1.0.0
fastapi==0.95.1
psycopg==3.1.8
uvicorn==0.21.1